requirements.txt
beautifulsoup4==4.13.4
Flask==3.1.1
langchain==0.3.26
langchain_community==0.3.26
langchain_core==0.3.67
langchain_experimental==0.3.4
langchain_groq==0.3.4
langchain_huggingface==0.3.0
python-dotenv==1.1.1
Requests==2.32.4
streamlit==1.46.1
torch==2.7.1
transformers==4.52.4
yt_dlp==2025.6.30
